Refactor "TRR test suite" to "TRR verification package"
